requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
decorator==4.4.2
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.24.4
Pillow==10.0.1
proglog==0.1.10
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
